Imp(?): default schema in app.prop instead of entity

+++ RR-Backend/src/main/java/com/scor/rr/domain/TargetBuild/Search/ShortCut.java
@@ -15,7 +15,7 @@ public class ShortCut {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer shortCutId;
+    private Long shortCutId;
 
     @Column(unique = true)
     private String shortCutLabel;
