+++ RR-Backend/src/main/java/com/scor/rr/repository/TargetBuild/UserTagRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserTagRepository extends JpaRepository<UserTag, Integer> {
+public interface UserTagRepository extends JpaRepository<UserTag, Long> {
     Optional<UserTag> findByTagIdAndUser(Long tagId, Integer userId);
 }
